refactor: Board 패키징
