+++ src/main/kotlin/omok/mapper/ArkMapper.kt
@@ -1,9 +1,9 @@
 package omok.mapper
 
-import omok.model.Board
 import omok.model.OmokStone
 import omok.model.Position
 import omok.model.StoneColor
+import omok.model.board.Board
 
 fun Board.toArkOmokBoard(): List<List<Int>> {
     val arkBoard = MutableList(15) { MutableList(15) { 0 } }
