+++ src/main/kotlin/omok/model/rule/OmokGameRule.kt
@@ -1,7 +1,7 @@
 package omok.model.rule
 
-import omok.model.Board
 import omok.model.OmokStone
+import omok.model.board.Board
 
 fun interface OmokGameRule {
     fun canPlaceStone(
